Continue refactoring, fix swiftlint
